move LocalIndex and all subclasses to separate dir
